fixes array count issues

+++ CLArrayType.h
@@ -20,7 +20,7 @@ typedef struct CLArrayType CLArrayType;
 CLArrayType *CLArrayTypeCreateWithObjects(void *, ...);
 CLArrayType *CLArrayTypeCreateWithList(void *, va_list, int);
 void *CLArrayTypeMutableCopy(CLArrayType *);
-int CLArrayTypeCount(CLArrayType *);
+size_t CLArrayTypeCount(CLArrayType *);
 void *CLArrayObjectAtIndex(CLArrayType *, int);
 bool CLArrayContainsObject(CLArrayType *, void *);
 CLArrayType *CLArrayCopy(CLArrayType *);
